+++ Sources/KinescopeSDK/Bags/KVO/KVOSubKey.swift
@@ -9,6 +9,8 @@ import Foundation
 
 enum KVOSubKey {
     case playerStatus
+    case playerItem
+    case playerItemBufferEmpty
     case playerItemStatus
     case playerTimeControlStatus
 }